Updated / Added new secure client server JS files
Worked with Dr. Newmann on getting tls connections to work between a
normal computer and levi through node. Basically what we had before but
now working and in less lines of code. So the new tls_client and server
are updated and working versions of the old levi_client and server.